Refactor Code
Related issues: BLD-1084
