+++ library/src/main/java/ai/bitlabs/sdk/views/WebActivity.kt
@@ -2,7 +2,7 @@ package ai.bitlabs.sdk.views
 
 import ai.bitlabs.sdk.BitLabs
 import ai.bitlabs.sdk.R
-import ai.bitlabs.sdk.setup
+import ai.bitlabs.sdk.util.setup
 import ai.bitlabs.sdk.util.BUNDLE_KEY_PARAMS
 import ai.bitlabs.sdk.util.TAG
 import android.net.Uri
