--- library/src/main/res/drawable/background_circle_regular.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <stroke
-        android:width="2dp"
-        android:color="@android:color/black" />
-
-    <corners android:radius="100dp" />
-</shape>
