+++ library/src/main/res/drawable/ic_circle_play_light.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M356.5,235.5C363.6,239.9 368,247.6 368,256C368,264.4 363.6,272.1 356.5,276.5L212.5,364.5C205.1,369 195.8,369.2 188.3,364.9C180.7,360.7 176,352.7 176,344V167.1C176,159.3 180.7,151.3 188.3,147.1C195.8,142.8 205.1,142.1 212.5,147.5L356.5,235.5zM208,182.3V329.7L328.7,255.1L208,182.3zM0,256C0,114.6 114.6,0 256,0C397.4,0 512,114.6 512,256C512,397.4 397.4,512 256,512C114.6,512 0,397.4 0,256zM256,480C379.7,480 480,379.7 480,256C480,132.3 379.7,32 256,32C132.3,32 32,132.3 32,256C32,379.7 132.3,480 256,480z"/>
+</vector>
