+++ library/src/main/res/layout/view_survey.xml
@@ -5,9 +5,9 @@
     android:id="@+id/fl_widget_container"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_marginHorizontal="8dp"
+    android:layout_marginHorizontal="4dp"
     android:background="@drawable/background_card_blue"
-    android:minWidth="225dp"
+    android:minWidth="300dp"
     android:padding="10dp">
 
     <LinearLayout
@@ -39,7 +39,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:indeterminateDrawable="@drawable/star_ratingbar"
-                android:maxHeight="25dip"
+                android:maxHeight="25dp"
                 android:minHeight="0dp"
                 android:numStars="5"
                 android:progressDrawable="@drawable/star_ratingbar"
@@ -69,13 +69,10 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="center"
-            android:layout_margin="8dp"
-            android:background="@drawable/background_circle_regular"
+            android:layout_marginStart="8dp"
             android:contentDescription="Start"
-            android:paddingHorizontal="10dp"
-            android:paddingVertical="8dp"
             android:scaleType="centerInside"
-            android:src="@drawable/ic_play_regular" />
+            android:src="@drawable/ic_circle_play_light" />
 
         <LinearLayout
             android:layout_width="wrap_content"
